requirements.txt
streamlit
torch
transformers
requests
Pillow
